/ redirects to /cards and displays data from controller
This is a very simple version of the flow I'd expect for the site: go to the
main page, send data to the controller to prepare the next page. In this case,
the main page isn't doing anything, and just redirecting to the second page,
whose controller has "card data" stubbed into it for the view.